refactor(map): Relocate iOS map utility functions
- Move `calculateDistance` from the deleted `MapUtils` to `MosqueClusterer`.
- Move `getClusterDistance` from the deleted `MapUtils` to `MapMarkerManager`.
- Extract `calculateZoomLevel` into a private function within `MapView.ios.kt`.
- Use named arguments for `CGRectMake`, `CGSizeMake`, and `MKCoordinateSpanMake` for improved readability in iOS map-related files.
